log added

+++ src/main/java/com/ecommerce/shopping/service/UserService.java
@@ -7,21 +7,30 @@ import com.ecommerce.shopping.model.Address;
 import com.ecommerce.shopping.model.User;
 import com.ecommerce.shopping.repository.UserRepository;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
-
+@Slf4j
 @Service
 @AllArgsConstructor
 public class UserService {
     private final UserRepository userRepository;
 
     public List<UserResponseDto> fetchAllUser(){
+
+        log.info("Fetching all Users from Database");
+
         List<User> users = userRepository.findAll();
+
+        log.info("Total users fetched: {}", users.size());
 
     return users.stream()
                 .map(user ->{Address address = user.getAddress();
+
+                    log.debug("Mapping user with email: {}", user.getEmail());
+
                 AddressDto addressDto = new AddressDto(
                         address.getStreet(),
                         address.getState(),
@@ -50,7 +59,9 @@ public class UserService {
         user.setPhoneNumber(userRequestDto.getPhoneNumber());
         user.setAddress(address);
 
+        log.info("Creating user with email: {}", user.getEmail());
         User savedUser = userRepository.save(user);
+        log.info("User added successfully in DataBase with ID: {}", savedUser.getId());
 
         AddressDto addressDto = new AddressDto(
                 savedUser.getAddress().getStreet(),
@@ -64,9 +75,14 @@ public class UserService {
         return new UserResponseDto(savedUser.getFirstName(), savedUser.getEmail(), addressDto);
     }
 
-    private  UserResponseDto getUserBYId(Long id, UserRequestDto userRequestDto) {
+    public UserResponseDto getUserBYId(Long id) {
 
-        User user = userRepository.findById(id).orElseThrow(()-> new RuntimeException("User Not found in this id:"+id));
+        log.info("Fetching user by ID: {}", id);
+
+        User user = userRepository.findById(id).orElseThrow(()->{
+            log.error("User not found with ID: {}", id);
+            return new RuntimeException("User not found");
+        });
 
         AddressDto addressDto = new AddressDto(
                 user.getAddress().getStreet(),
@@ -75,12 +91,18 @@ public class UserService {
                 user.getAddress().getZipcode()
         );
 
+        log.info("User fetched successfully with ID: {}", id);
         return new UserResponseDto(user.getFirstName(), user.getEmail(), addressDto);
     }
 
 
     public UserResponseDto updateUser(Long id, UserRequestDto userRequestDto) {
-        User updatedUser = userRepository.findById(id).orElseThrow(()-> new RuntimeException("User Not Found in this id:"+id));
+
+        log.info("Fetching the existing user data from id: {}", id);
+        User updatedUser = userRepository.findById(id).orElseThrow(()-> {
+            log.error("User not found with ID: {}", id);
+        return new RuntimeException("User not found");
+        });
 
         Address address = new Address();
         address.setState(userRequestDto.getAddress().getState());
@@ -88,6 +110,7 @@ public class UserService {
         address.setCountry(userRequestDto.getAddress().getCountry());
         address.setZipcode(userRequestDto.getAddress().getZipcode());
 
+        log.info("Updating old user with new data for id: {}", id);
         updatedUser.setFirstName(userRequestDto.getFirstName());
         updatedUser.setLastName(userRequestDto.getLastName());
         updatedUser.setEmail(userRequestDto.getEmail());
@@ -102,6 +125,8 @@ public class UserService {
         );
 
         User savedUser = userRepository.save(updatedUser);
+        log.info("User updated successfully for ID: {}", id);
+        log.info("Updated user stored successfully in database with ID: {}", savedUser.getId());
 
         return new UserResponseDto(savedUser.getFirstName(),
                 savedUser.getEmail(),
@@ -111,7 +136,13 @@ public class UserService {
 
 
     public String deleteUserById(Long id) {
-        userRepository.deleteById(id);
-        return "User Deleted Successfully from id:" + id;
+        if (!userRepository.existsById(id)){
+            log.error("User not found with ID: {}", id);
+            throw new RuntimeException("User not found with id"+id);        }
+        else {
+            userRepository.deleteById(id);
+            log.info("User deleted successfully with ID: {}", id);
+            return "User Deleted Successfully from id:" + id;
+        }
     }
 }
